pyproject.toml
[tool.poetry]
name = "cw7"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.6.0"
celery = "^5.4.0"
redis = "^5.2.1"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.9"
coverage = "^7.6.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
